pyproject.toml
[project]
name = "assistant-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.25.15",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langsmith>=0.7.22",
    "python-dotenv>=1.2.2",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
